requirements.txt
streamlit
langchain
langchain_google_genai
google-generativeai
python-dotenv